# Optimization: replaces lodash's memoize with a custom, much faster one