--- a/lib/kuzzle.js
+++ b/lib/kuzzle.js
@@ -52,6 +52,7 @@ const StorageEngine = require('./core/storage/storageEngine');
 const runShutdown = require('./util/shutdown');
 const Logger = require('./util/log');
 const Promback = require('./util/promback');
+const memoize = require('./util/memoize');
 
 
 /**
@@ -85,7 +86,7 @@ function pipeCallback (error, updated) {
  * @param {String} event
  * @returns {Array<String>} wildcard events
  */
-const getWildcardEvents = _.memoize(event => {
+const getWildcardEvents = memoize(event => {
   const
     events = [event],
     delimIndex = event.lastIndexOf(':');