--- a/lib/util/memoize.js
+++ b/lib/util/memoize.js
@@ -1,0 +1,47 @@
+/*
+ * Kuzzle, a backend software, self-hostable and ready to use
+ * to power modern apps
+ *
+ * Copyright 2015-2020 Kuzzle
+ * mailto: support AT kuzzle.io
+ * website: http://kuzzle.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use strict';
+
+function memoize (fn, resolver = null) {
+  const memoized = (...args) => {
+    const key = memoized.resolver !== null
+      ? memoized.resolver(args)
+      : args[0];
+
+    let result = memoized.cache.get(key);
+
+    if (result === undefined) {
+      result = memoized.fn(...args);
+      memoized.cache.set(key, result);
+    }
+
+    return result;
+  };
+
+  memoized.fn = fn;
+  memoized.cache = new Map();
+  memoized.resolver = resolver;
+
+  return memoized;
+}
+
+module.exports = memoize;