# Description

We use [memoization](https://en.wikipedia.org/wiki/Memoization) to skip calculation on one function in a critical section of our code (note for later: this should definitly be used at more than one place).

Until now we were using lodash's [memoize](https://lodash.com/docs/4.17.15#memoize). This PR replaces it with a (much) faster version of it.

# Benchmarks

```
custom memoize x 96,849,274 ops/sec ±0.65% (96 runs sampled)
custom memoize with resolver x 107,643,904 ops/sec ±0.93% (94 runs sampled)
_.memoize x 24,907,495 ops/sec ±0.86% (94 runs sampled)
_.memoize with resolver x 18,251,824 ops/sec ±0.54% (94 runs sampled)
```